Revert "prefer using cython unless there's no .pyx but a .cpp"
This reverts commit 52cce6fb8fe7d3149fe15b2ac99e5ad529fc9c87.

+++ setup.py
@@ -61,15 +61,15 @@ else:  # set `__version__` from _release.py:
 package_include = os.path.join(pkg_name, 'include')
 
 _src = {ext: _path_under_setup(pkg_name, '_cvodes.' + ext) for ext in "cpp pyx".split()}
-if os.path.exists(_src["pyx"]):
+if _HAVE_CYTHON and os.path.exists(_src["pyx"]):
+    # Possible that a new release of Python needs a re-rendered Cython source,
+    # or that we want to include possible bug-fix to Cython, disable by manually
+    # deleting .pyx file from source distribution.
     USE_CYTHON = True
     if os.path.exists(_src['cpp']):
         os.unlink(_src['cpp'])  # ensure c++ source is re-generated.
-    USE_CYTHON = False
-elif os.path.exists(_src['cpp']):
-    USE_CYTHON = False
 else:
-    raise OSError("Could not find neither .cpp nor .pyx source file")
+    USE_CYTHON = False
 
 ext_modules = []
 env = dict(
